update /requests and /emails according to API doc
