+++ backend/kafka_init.sh
@@ -8,5 +8,5 @@ bootstrap_server="kafka:29092"
 
 for topic in "${topics[@]}"; do
   # Create each topic using kafka-topics.sh
-  docker exec -d kafka kafka-topics.sh --create --topic "$topic" --partitions 1 --replication-factor 1 --bootstrap-server "$bootstrap_server"
+  docker exec -it kafka kafka-topics.sh --create --topic "$topic" --partitions 1 --replication-factor 1 --bootstrap-server "$bootstrap_server"
 done
