+++ docker-compose.yaml
@@ -136,6 +136,7 @@ services:
       - KAFKA_CFG_NODE_ID=0
       - KAFKA_CFG_PROCESS_ROLES=controller,broker
       - KAFKA_CFG_CONTROLLER_QUORUM_VOTERS=0@kafka:9093
+      - KAFKA_CFG_AUTO_CREATE_TOPICS_ENABLE=false
       # Listeners
       - KAFKA_CFG_LISTENERS=LISTENER_EXTERNAL://kafka:9092,LISTENER_INTERNAL://kafka:29092,CONTROLLER://:9093
       - KAFKA_CFG_ADVERTISED_LISTENERS=LISTENER_EXTERNAL://localhost:9092,LISTENER_INTERNAL://kafka:29092
