+++ email/endpoint.py
@@ -10,7 +10,7 @@ def get_emails(user_config, n_mails: int) -> dict:
     clean_mails = []
     for email_id, raw_texts in raw_emails:
         plain_text = clean_email_from_raw_texts(raw_texts)
-        email = {"email_id": email_id, "address": user_config['username'], "item": plain_text}
+        email = {"email_id": email_id, "item": plain_text}
         clean_mails.append(email)
         logger.debug(f"{email_id}:\n{plain_text}\n")
     emails = {"items": clean_mails}
@@ -42,7 +42,7 @@ async def aget_emails(user_config, n_mails: int) -> dict:
     clean_mails = []
     for email_id, raw_texts in raw_emails:
         plain_text = clean_email_from_raw_texts(raw_texts)
-        email = {"email_id": email_id, "address": user_config['username'], "item": plain_text}
+        email = {"email_id": email_id, "item": plain_text}
         clean_mails.append(email)
         logger.debug(f"{email_id}:\n{plain_text}\n")
     emails = {"items": clean_mails}
